Moved the http middleware used for config endpoint in the auth package.
